pyproject.toml
[project]
name = "indices_precio_globales"
version = "0.1.0"
description = "Analisis de indices de precios globales en Republica Dominicana"
authors = [
    { name = "Mariana Rodriguez" }
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "openpyxl>=3.0.0",
    "pandera>=0.16.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "matplotlib>=3.0.0",
    "pytest>=6.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
exclude = [
    "Notebook",
    ".ipynb_checkpoints"
]